Streaming store-gateway: Move seriesHasher
